@@ -1,11 +1,11 @@
-// +build tcell windows
+//go:build tcell || windows
 
 package tui
 
 import (
 	"testing"
 
-	"github.com/gdamore/tcell"
+	"github.com/gdamore/tcell/v2"
 	"github.com/junegunn/fzf/src/util"
 )
 
