@@ -1,4 +1,4 @@
-// +build windows
+//go:build windows
 
 package util
 
@@ -80,4 +80,8 @@ func SetNonblock(file *os.File, nonblock bool) {
 // Read executes syscall.Read on file descriptor
 func Read(fd int, b []byte) (int, error) {
 	return syscall.Read(syscall.Handle(fd), b)
+}
+
+func SetStdin(file *os.File) {
+	// No-op
 }
