@@ -1,9 +1,9 @@
 (defproject vim-clojure-static "0.1.0"
-  :description "Utilities and tests for Vim's Clojure runtime files."
-  :url "https://github.com/clojure-vim/clojure.vim"
-  :license {:name "Vim License"
-            :url "http://vimdoc.sourceforge.net/htmldoc/uganda.html#license"
-            :comments ":help license"}
-  :dependencies [[org.clojure/data.csv "1.0.0"]
-                 [org.clojure/clojure "1.10.3"]
+  :description  "Utilities and tests for Vim's Clojure runtime files."
+  :url          "https://github.com/clojure-vim/clojure.vim"
+  :license      {:name     "Vim License"
+                 :url      "http://vimdoc.sourceforge.net/htmldoc/uganda.html#license"
+                 :comments ":help license"}
+  :dependencies [[org.clojure/clojure "1.11.1"]
+                 [org.clojure/data.csv "1.0.1"]
                  [frak "0.1.9"]])
