@@ -90,6 +90,9 @@ const (
 
 	Change
 	BackwardEOF
+	Start
+	Load
+	Focus
 
 	AltBS
 
@@ -266,7 +269,11 @@ type ColorTheme struct {
 	Cursor       ColorAttr
 	Selected     ColorAttr
 	Header       ColorAttr
+	Separator    ColorAttr
+	Scrollbar    ColorAttr
 	Border       ColorAttr
+	BorderLabel  ColorAttr
+	PreviewLabel ColorAttr
 }
 
 type Event struct {
@@ -291,6 +298,8 @@ const (
 	BorderNone BorderShape = iota
 	BorderRounded
 	BorderSharp
+	BorderBold
+	BorderDouble
 	BorderHorizontal
 	BorderVertical
 	BorderTop
@@ -298,6 +307,22 @@ const (
 	BorderLeft
 	BorderRight
 )
+
+func (s BorderShape) HasRight() bool {
+	switch s {
+	case BorderNone, BorderLeft, BorderTop, BorderBottom, BorderHorizontal: // No right
+		return false
+	}
+	return true
+}
+
+func (s BorderShape) HasTop() bool {
+	switch s {
+	case BorderNone, BorderLeft, BorderRight, BorderBottom, BorderVertical: // No top
+		return false
+	}
+	return true
+}
 
 type BorderStyle struct {
 	shape       BorderShape
@@ -312,18 +337,19 @@ type BorderStyle struct {
 type BorderCharacter int
 
 func MakeBorderStyle(shape BorderShape, unicode bool) BorderStyle {
-	if unicode {
-		if shape == BorderRounded {
-			return BorderStyle{
-				shape:       shape,
-				horizontal:  '─',
-				vertical:    '│',
-				topLeft:     '╭',
-				topRight:    '╮',
-				bottomLeft:  '╰',
-				bottomRight: '╯',
-			}
+	if !unicode {
+		return BorderStyle{
+			shape:       shape,
+			horizontal:  '-',
+			vertical:    '|',
+			topLeft:     '+',
+			topRight:    '+',
+			bottomLeft:  '+',
+			bottomRight: '+',
 		}
+	}
+	switch shape {
+	case BorderSharp:
 		return BorderStyle{
 			shape:       shape,
 			horizontal:  '─',
@@ -333,15 +359,35 @@ func MakeBorderStyle(shape BorderShape, unicode bool) BorderStyle {
 			bottomLeft:  '└',
 			bottomRight: '┘',
 		}
+	case BorderBold:
+		return BorderStyle{
+			shape:       shape,
+			horizontal:  '━',
+			vertical:    '┃',
+			topLeft:     '┏',
+			topRight:    '┓',
+			bottomLeft:  '┗',
+			bottomRight: '┛',
+		}
+	case BorderDouble:
+		return BorderStyle{
+			shape:       shape,
+			horizontal:  '═',
+			vertical:    '║',
+			topLeft:     '╔',
+			topRight:    '╗',
+			bottomLeft:  '╚',
+			bottomRight: '╝',
+		}
 	}
 	return BorderStyle{
 		shape:       shape,
-		horizontal:  '-',
-		vertical:    '|',
-		topLeft:     '+',
-		topRight:    '+',
-		bottomLeft:  '+',
-		bottomRight: '+',
+		horizontal:  '─',
+		vertical:    '│',
+		topLeft:     '╭',
+		topRight:    '╮',
+		bottomLeft:  '╰',
+		bottomRight: '╯',
 	}
 }
 
@@ -358,12 +404,14 @@ func MakeTransparentBorder() BorderStyle {
 
 type Renderer interface {
 	Init()
+	Resize(maxHeightFunc func(int) int)
 	Pause(clear bool)
 	Resume(clear bool, sigcont bool)
 	Clear()
 	RefreshWindows(windows []Window)
 	Refresh()
 	Close()
+	NeedScrollbarRedraw() bool
 
 	GetChar() Event
 
@@ -379,6 +427,7 @@ type Window interface {
 	Width() int
 	Height() int
 
+	DrawHBorder()
 	Refresh()
 	FinishFill()
 	Close()
@@ -401,7 +450,7 @@ type FullscreenRenderer struct {
 	mouse        bool
 	forceBlack   bool
 	prevDownTime time.Time
-	clickY       []int
+	clicks       [][2]int
 }
 
 func NewFullscreenRenderer(theme *ColorTheme, forceBlack bool, mouse bool) Renderer {
@@ -410,7 +459,7 @@ func NewFullscreenRenderer(theme *ColorTheme, forceBlack bool, mouse bool) Rende
 		mouse:        mouse,
 		forceBlack:   forceBlack,
 		prevDownTime: time.Unix(0, 0),
-		clickY:       []int{}}
+		clicks:       [][2]int{}}
 	return r
 }
 
@@ -436,22 +485,22 @@ var (
 	ColSpinner              ColorPair
 	ColInfo                 ColorPair
 	ColHeader               ColorPair
+	ColSeparator            ColorPair
+	ColScrollbar            ColorPair
 	ColBorder               ColorPair
 	ColPreview              ColorPair
 	ColPreviewBorder        ColorPair
+	ColBorderLabel          ColorPair
+	ColPreviewLabel         ColorPair
 )
 
 func EmptyTheme() *ColorTheme {
 	return &ColorTheme{
 		Colored:      true,
 		Input:        ColorAttr{colUndefined, AttrUndefined},
-		Disabled:     ColorAttr{colUndefined, AttrUndefined},
 		Fg:           ColorAttr{colUndefined, AttrUndefined},
 		Bg:           ColorAttr{colUndefined, AttrUndefined},
-		PreviewFg:    ColorAttr{colUndefined, AttrUndefined},
-		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
 		DarkBg:       ColorAttr{colUndefined, AttrUndefined},
-		Gutter:       ColorAttr{colUndefined, AttrUndefined},
 		Prompt:       ColorAttr{colUndefined, AttrUndefined},
 		Match:        ColorAttr{colUndefined, AttrUndefined},
 		Current:      ColorAttr{colUndefined, AttrUndefined},
@@ -461,20 +510,25 @@ func EmptyTheme() *ColorTheme {
 		Cursor:       ColorAttr{colUndefined, AttrUndefined},
 		Selected:     ColorAttr{colUndefined, AttrUndefined},
 		Header:       ColorAttr{colUndefined, AttrUndefined},
-		Border:       ColorAttr{colUndefined, AttrUndefined}}
+		Border:       ColorAttr{colUndefined, AttrUndefined},
+		BorderLabel:  ColorAttr{colUndefined, AttrUndefined},
+		Disabled:     ColorAttr{colUndefined, AttrUndefined},
+		PreviewFg:    ColorAttr{colUndefined, AttrUndefined},
+		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
+		Gutter:       ColorAttr{colUndefined, AttrUndefined},
+		PreviewLabel: ColorAttr{colUndefined, AttrUndefined},
+		Separator:    ColorAttr{colUndefined, AttrUndefined},
+		Scrollbar:    ColorAttr{colUndefined, AttrUndefined},
+	}
 }
 
 func NoColorTheme() *ColorTheme {
 	return &ColorTheme{
 		Colored:      false,
 		Input:        ColorAttr{colDefault, AttrRegular},
-		Disabled:     ColorAttr{colDefault, AttrRegular},
 		Fg:           ColorAttr{colDefault, AttrRegular},
 		Bg:           ColorAttr{colDefault, AttrRegular},
-		PreviewFg:    ColorAttr{colDefault, AttrRegular},
-		PreviewBg:    ColorAttr{colDefault, AttrRegular},
 		DarkBg:       ColorAttr{colDefault, AttrRegular},
-		Gutter:       ColorAttr{colDefault, AttrRegular},
 		Prompt:       ColorAttr{colDefault, AttrRegular},
 		Match:        ColorAttr{colDefault, Underline},
 		Current:      ColorAttr{colDefault, Reverse},
@@ -484,7 +538,16 @@ func NoColorTheme() *ColorTheme {
 		Cursor:       ColorAttr{colDefault, AttrRegular},
 		Selected:     ColorAttr{colDefault, AttrRegular},
 		Header:       ColorAttr{colDefault, AttrRegular},
-		Border:       ColorAttr{colDefault, AttrRegular}}
+		Border:       ColorAttr{colDefault, AttrRegular},
+		BorderLabel:  ColorAttr{colDefault, AttrRegular},
+		Disabled:     ColorAttr{colDefault, AttrRegular},
+		PreviewFg:    ColorAttr{colDefault, AttrRegular},
+		PreviewBg:    ColorAttr{colDefault, AttrRegular},
+		Gutter:       ColorAttr{colDefault, AttrRegular},
+		PreviewLabel: ColorAttr{colDefault, AttrRegular},
+		Separator:    ColorAttr{colDefault, AttrRegular},
+		Scrollbar:    ColorAttr{colDefault, AttrRegular},
+	}
 }
 
 func errorExit(message string) {
@@ -496,13 +559,9 @@ func init() {
 	Default16 = &ColorTheme{
 		Colored:      true,
 		Input:        ColorAttr{colDefault, AttrUndefined},
-		Disabled:     ColorAttr{colUndefined, AttrUndefined},
 		Fg:           ColorAttr{colDefault, AttrUndefined},
 		Bg:           ColorAttr{colDefault, AttrUndefined},
-		PreviewFg:    ColorAttr{colUndefined, AttrUndefined},
-		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
 		DarkBg:       ColorAttr{colBlack, AttrUndefined},
-		Gutter:       ColorAttr{colUndefined, AttrUndefined},
 		Prompt:       ColorAttr{colBlue, AttrUndefined},
 		Match:        ColorAttr{colGreen, AttrUndefined},
 		Current:      ColorAttr{colYellow, AttrUndefined},
@@ -512,17 +571,22 @@ func init() {
 		Cursor:       ColorAttr{colRed, AttrUndefined},
 		Selected:     ColorAttr{colMagenta, AttrUndefined},
 		Header:       ColorAttr{colCyan, AttrUndefined},
-		Border:       ColorAttr{colBlack, AttrUndefined}}
+		Border:       ColorAttr{colBlack, AttrUndefined},
+		BorderLabel:  ColorAttr{colWhite, AttrUndefined},
+		Disabled:     ColorAttr{colUndefined, AttrUndefined},
+		PreviewFg:    ColorAttr{colUndefined, AttrUndefined},
+		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
+		Gutter:       ColorAttr{colUndefined, AttrUndefined},
+		PreviewLabel: ColorAttr{colUndefined, AttrUndefined},
+		Separator:    ColorAttr{colUndefined, AttrUndefined},
+		Scrollbar:    ColorAttr{colUndefined, AttrUndefined},
+	}
 	Dark256 = &ColorTheme{
 		Colored:      true,
 		Input:        ColorAttr{colDefault, AttrUndefined},
-		Disabled:     ColorAttr{colUndefined, AttrUndefined},
 		Fg:           ColorAttr{colDefault, AttrUndefined},
 		Bg:           ColorAttr{colDefault, AttrUndefined},
-		PreviewFg:    ColorAttr{colUndefined, AttrUndefined},
-		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
 		DarkBg:       ColorAttr{236, AttrUndefined},
-		Gutter:       ColorAttr{colUndefined, AttrUndefined},
 		Prompt:       ColorAttr{110, AttrUndefined},
 		Match:        ColorAttr{108, AttrUndefined},
 		Current:      ColorAttr{254, AttrUndefined},
@@ -532,17 +596,22 @@ func init() {
 		Cursor:       ColorAttr{161, AttrUndefined},
 		Selected:     ColorAttr{168, AttrUndefined},
 		Header:       ColorAttr{109, AttrUndefined},
-		Border:       ColorAttr{59, AttrUndefined}}
+		Border:       ColorAttr{59, AttrUndefined},
+		BorderLabel:  ColorAttr{145, AttrUndefined},
+		Disabled:     ColorAttr{colUndefined, AttrUndefined},
+		PreviewFg:    ColorAttr{colUndefined, AttrUndefined},
+		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
+		Gutter:       ColorAttr{colUndefined, AttrUndefined},
+		PreviewLabel: ColorAttr{colUndefined, AttrUndefined},
+		Separator:    ColorAttr{colUndefined, AttrUndefined},
+		Scrollbar:    ColorAttr{colUndefined, AttrUndefined},
+	}
 	Light256 = &ColorTheme{
 		Colored:      true,
 		Input:        ColorAttr{colDefault, AttrUndefined},
-		Disabled:     ColorAttr{colUndefined, AttrUndefined},
 		Fg:           ColorAttr{colDefault, AttrUndefined},
 		Bg:           ColorAttr{colDefault, AttrUndefined},
-		PreviewFg:    ColorAttr{colUndefined, AttrUndefined},
-		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
 		DarkBg:       ColorAttr{251, AttrUndefined},
-		Gutter:       ColorAttr{colUndefined, AttrUndefined},
 		Prompt:       ColorAttr{25, AttrUndefined},
 		Match:        ColorAttr{66, AttrUndefined},
 		Current:      ColorAttr{237, AttrUndefined},
@@ -552,7 +621,16 @@ func init() {
 		Cursor:       ColorAttr{161, AttrUndefined},
 		Selected:     ColorAttr{168, AttrUndefined},
 		Header:       ColorAttr{31, AttrUndefined},
-		Border:       ColorAttr{145, AttrUndefined}}
+		Border:       ColorAttr{145, AttrUndefined},
+		BorderLabel:  ColorAttr{59, AttrUndefined},
+		Disabled:     ColorAttr{colUndefined, AttrUndefined},
+		PreviewFg:    ColorAttr{colUndefined, AttrUndefined},
+		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
+		Gutter:       ColorAttr{colUndefined, AttrUndefined},
+		PreviewLabel: ColorAttr{colUndefined, AttrUndefined},
+		Separator:    ColorAttr{colUndefined, AttrUndefined},
+		Scrollbar:    ColorAttr{colUndefined, AttrUndefined},
+	}
 }
 
 func initTheme(theme *ColorTheme, baseTheme *ColorTheme, forceBlack bool) {
@@ -571,13 +649,9 @@ func initTheme(theme *ColorTheme, baseTheme *ColorTheme, forceBlack bool) {
 		return c
 	}
 	theme.Input = o(baseTheme.Input, theme.Input)
-	theme.Disabled = o(theme.Input, o(baseTheme.Disabled, theme.Disabled))
 	theme.Fg = o(baseTheme.Fg, theme.Fg)
 	theme.Bg = o(baseTheme.Bg, theme.Bg)
-	theme.PreviewFg = o(theme.Fg, o(baseTheme.PreviewFg, theme.PreviewFg))
-	theme.PreviewBg = o(theme.Bg, o(baseTheme.PreviewBg, theme.PreviewBg))
 	theme.DarkBg = o(baseTheme.DarkBg, theme.DarkBg)
-	theme.Gutter = o(theme.DarkBg, o(baseTheme.Gutter, theme.Gutter))
 	theme.Prompt = o(baseTheme.Prompt, theme.Prompt)
 	theme.Match = o(baseTheme.Match, theme.Match)
 	theme.Current = o(baseTheme.Current, theme.Current)
@@ -588,6 +662,16 @@ func initTheme(theme *ColorTheme, baseTheme *ColorTheme, forceBlack bool) {
 	theme.Selected = o(baseTheme.Selected, theme.Selected)
 	theme.Header = o(baseTheme.Header, theme.Header)
 	theme.Border = o(baseTheme.Border, theme.Border)
+	theme.BorderLabel = o(baseTheme.BorderLabel, theme.BorderLabel)
+
+	// These colors are not defined in the base themes
+	theme.Disabled = o(theme.Input, theme.Disabled)
+	theme.Gutter = o(theme.DarkBg, theme.Gutter)
+	theme.PreviewFg = o(theme.Fg, theme.PreviewFg)
+	theme.PreviewBg = o(theme.Bg, theme.PreviewBg)
+	theme.PreviewLabel = o(theme.BorderLabel, theme.PreviewLabel)
+	theme.Separator = o(theme.Border, theme.Separator)
+	theme.Scrollbar = o(theme.Border, theme.Scrollbar)
 
 	initPalette(theme)
 }
@@ -619,7 +703,11 @@ func initPalette(theme *ColorTheme) {
 	ColSpinner = pair(theme.Spinner, theme.Bg)
 	ColInfo = pair(theme.Info, theme.Bg)
 	ColHeader = pair(theme.Header, theme.Bg)
+	ColSeparator = pair(theme.Separator, theme.Bg)
+	ColScrollbar = pair(theme.Scrollbar, theme.Bg)
 	ColBorder = pair(theme.Border, theme.Bg)
+	ColBorderLabel = pair(theme.BorderLabel, theme.Bg)
+	ColPreviewLabel = pair(theme.PreviewLabel, theme.PreviewBg)
 	ColPreview = pair(theme.PreviewFg, theme.PreviewBg)
 	ColPreviewBorder = pair(theme.Border, theme.PreviewBg)
 }
